+++ src/test/groovy/org/gmetrics/metric/cyclomatic/CyclomaticComplexityMetric_MethodTest.groovy
@@ -255,4 +255,31 @@ class CyclomaticComplexityMetric_MethodTest extends AbstractMetricTestCase {
         assert calculateForMethod(SOURCE) == 9
     }
 
+    @Test
+    void testCalculate_CorrectLineNumberForAnnotatedMethod_NextLine() {
+        final SOURCE = """
+            @SomeAnnotation
+            def myMethod() { }
+        """
+        assert metricLineNumber(findFirstMethod(SOURCE)) == 3
+    }
+
+    @Test
+    void testCalculate_CorrectLineNumberForAnnotatedMethod_SameLine() {
+        final SOURCE = """
+            @SomeAnnotation def myMethod() { }
+        """
+        assert metricLineNumber(findFirstMethod(SOURCE)) == 2
+    }
+
+    @Test
+    void testCalculate_CorrectLineNumberForAnnotatedMethod_CommentInBetween() {
+        final SOURCE = """
+            @SomeAnnotation
+            // comment
+            def myMethod() { }
+        """
+        assert metricLineNumber(findFirstMethod(SOURCE)) == 4
+    }
+
 }